Revert "Feat/testes entidades capitulos"
